refactor(styles): move style module into app/styles/ and drop empty globals.css
