+++ app/address/[address]/styles.css
@@ -1,1 +1,1 @@
-@import '../../styles.css';
+@import '../../styles/styles.css';
